rename mensa_manager to mensa_event
